Cleanup main code
