+++ Sources/SwiftRetrier/Core/Model/AttemptFailure.swift
