+++ Sources/SwiftRetrier/Core/Model/RetrierEvent.swift
