--- Sources/SwiftRetrier/Core/Model/RetryError.swift
@@ -1,3 +1,0 @@
-public enum RetryError: Error {
-    case conditionPublisherCompleted
-}
