+++ Sources/SwiftRetrier/DSL/BuilderEntryPoints.swift
@@ -4,20 +4,20 @@ public func withExponentialBackoff(
     timeSlot: TimeInterval = ExponentialBackoffConstants.defaultTimeSlot,
     maxDelay: TimeInterval = ExponentialBackoffConstants.defaultMaxDelay,
     jitter: ExponentialBackoffRetryPolicy.Jitter = ExponentialBackoffConstants.defaultJitter
-) -> ColdRetrier {
+) -> Retrier {
     let policy = ExponentialBackoffRetryPolicy(timeSlot: timeSlot,
                                                maxDelay: maxDelay,
                                                jitter: jitter)
-    return ColdRetrier(policy: policy, conditionPublisher: nil)
+    return Retrier(policy: policy, conditionPublisher: nil)
 }
 
 public func withConstantDelay(
     _ delay: TimeInterval = ConstantDelayConstants.defaultDelay
-) -> ColdRetrier {
+) -> Retrier {
     let policy = ConstantDelayRetryPolicy(delay: delay)
-    return ColdRetrier(policy: policy, conditionPublisher: nil)
+    return Retrier(policy: policy, conditionPublisher: nil)
 }
 
-public func withNoDelay() -> ColdRetrier {
+public func withNoDelay() -> Retrier {
     withConstantDelay(0)
 }
