--- Sources/SwiftRetrier/Util/OnMain.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-func onMain(_ block: @escaping @Sendable () -> Void) {
-    if Thread.isMainThread {
-        block()
-    } else {
-        DispatchQueue.main.async(execute: block)
-    }
-}
