+++ Sources/SwiftRetrier/Util/Publisher+AsyncMapLatest.swift
@@ -1,8 +1,8 @@
 import Foundation
-import Combine
+@preconcurrency import Combine
 
 /// Adapted from : https://www.swiftbysundell.com/articles/calling-async-functions-within-a-combine-pipeline/
-/// to handle cancellation properly
+/// to handle cancellation properly and be compatible with Swift 6 mode
 public extension Publisher where Output: Sendable {
 
     /// Map output using an async closure.
@@ -17,18 +17,17 @@ public extension Publisher where Output: Sendable {
     ) -> AnyPublisher<T, Error> {
         mapError { $0 as Error }
             .map { value in
-                var task: Task<Void, Never>?
-                return Future<T, Error> { (promise: @escaping (Result<T, Error>) -> Void) in
-                    nonisolated(unsafe) let promise = promise
-                    task = Task { @MainActor [promise] in
-                        do {
-                            let output = try await transform(value)
-                            promise(.success(output))
-                        } catch {
-                            promise(.failure(error))
-                        }
+                let subject = PassthroughSubject<T, Error>()
+                let task = Task { @MainActor [subject] in
+                    do {
+                        let output = try await transform(value)
+                        subject.send(output)
+                        subject.send(completion: .finished)
+                    } catch {
+                        subject.send(completion: .failure(error))
                     }
-                }.handleEvents(receiveCancel: { task?.cancel() })
+                }
+                return subject.handleEvents(receiveCancel: { task.cancel() })
             }
             .switchToLatest()
             .eraseToAnyPublisher()
